requirements.txt
discord.py>=2.3.0

python-dotenv>=1.0.0

aiohttp>=3.9.0
#For Python 13.1 >
audioop-lts